requirements.txt
fastapi==0.103.0
uvicorn[standard]==0.23.0
mangum==0.17.0
openai==1.51.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
httpx==0.25.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
Pillow==11.0.0
PyPDF2==3.0.1
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.1.0
pytz==2024.1

# Supabase Auth integration
supabase==2.11.0
gotrue==2.10.1

# Database dependencies
sqlalchemy[asyncio]==2.0.35
asyncpg==0.31.0
alembic==1.13.1
python-dateutil==2.8.2

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
aiosqlite==0.19.0